1.4.1: 修复bug #10633, 社区活动参与者列表，删除人员时. 连续快点崩溃

+++ app/src/main/java/lilun/com/pensionlife/module/adapter/PartnersAdapter.java
@@ -47,7 +47,7 @@ public class PartnersAdapter extends QuickAdapter<Account> {
         if (!selectedList.contains(pos))
             selectedList.add(pos);
         else
-            selectedList.remove(pos);
+            selectedList.remove((Integer) pos);
     }
 
     public ArrayList<Integer> getSelectedList() {
